ai-agent/pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "A ReAct + RAG AI agent powered by **GPT-4o-mini** and **ChromaDB** for internal knowledge retrieval from Markdown documents."
authors = [
    {name = "Hiro",email = "webmaster@supaahiro.org"}
]
license = {text = "**No-Commercial License** (see LICENSE file)"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.33,<0.4.0)",
    "langchain-community (>=0.3.29,<0.4.0)",
    "langgraph (>=0.6.7,<0.7.0)",
    "langchain-chroma (>=0.2.6,<0.3.0)",
    "numexpr (>=2.12.1,<3.0.0)",
    "grandalf (>=0.8,<0.9)",
    "uv (>=0.8.18,<0.9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
